chore: moving under pkg umbrella
