+++ pkg/analyses/npm/matcher_test.go
