+++ pkg/analyses/npm/testdata/package.json
