+++ pkg/analyses/summary/helper_test.go
