+++ pkg/analyses/verbose/helper_test.go
